main

+++ index.js
@@ -4,8 +4,6 @@ import cors from "cors";
 import { MongoClient, ObjectId } from "mongodb";
 import bcrypt from "bcrypt";
 import jsonwebtocken from "jsonwebtoken";
-import { request } from "https";
-import { response } from "express";
 import {
   getAllUsers,
   modifyData,
@@ -13,6 +11,7 @@ import {
   userRegistration,
   findOneData,
 } from "./helper.js";
+// import { users } from "./routes/users.js";
 
 dotenv.config();
 
@@ -20,32 +19,22 @@ dotenv.config();
 
 const app = express();
 
-// ? CUSTOM MIDDLEWARE
-
-const authorisation = async (request, response, next) => {
-  const token = request.header("token");
-
-  const verifyToken = jsonwebtocken.verify(token, process.env.privateKey);
-  const inputData = { email: verifyToken.email };
-  const checkUserExisting = await findOneData(inputData);
-
-  if (verifyToken && checkUserExisting) {
-    next();
-  } else {
-    response.status(401).send("Un Authorised");
-  }
-};
-
 app.use(express.json());
 app.use(cors());
 
 const PORT = 7000;
 
-const MONGO_URL = process.env.MONGO_URL;
+export const MONGO_URL = process.env.MONGO_URL;
 
 export const client = await createConnection();
 
-// ! LOGIN VERIFICATION
+// ! INITIAL ROUTE
+
+app.get("/", (req, res) => {
+  res.send("Deployment Success");
+});
+
+//! LOGIN VERIFICATION
 
 app.post("/login", async (request, response) => {
   const { phone, password } = request.body;
@@ -62,16 +51,12 @@ app.post("/login", async (request, response) => {
     } else {
       const token = await tokenGenerator(signIn.email);
       response.send({
-        message: `Login Successful`,
+        message: { name: signIn.name, email: signIn.email },
         token: token,
         status: "successful",
       });
     }
   }
-});
-
-app.get("/", (req, res) => {
-  res.send("Deployment Success");
 });
 
 // ? SIGN UP USER
@@ -104,46 +89,46 @@ app.post("/create/user", async (req, res) => {
 
 app.post("/createUser/byAdmin", async (req, res) => {
   const { name, contact, dob, email } = req.body;
+  const { token } = req.headers;
 
-  const password = "123456789";
-  const conformPassword = "123456789";
+  const crackData = await jsonwebtocken.verify(token, process.env.privateKey);
+  const mail = crackData.email;
+  const inputData = { email: mail };
+  const checkUserExists = await findOneData(inputData);
+  if (!checkUserExists) {
+    response.status(401).send("Unauthorized user");
+  } else {
+    const hashPassword = await createPassword("12345678");
 
-  const hashPassword = await createPassword("12345678");
+    const hashConformPassword = await createPassword("12345678");
 
-  const hashConformPassword = await createPassword("12345678");
+    try {
+      const signUpData = {
+        name,
+        contact,
+        dob,
+        email,
+        hashPassword,
+        hashConformPassword,
+      };
 
-  try {
-    const signUpData = {
-      name,
-      contact,
-      dob,
-      email,
-      hashPassword,
-      hashConformPassword,
-    };
-
-    if (password !== conformPassword) {
-      res.status("401").send("Password not Matching");
-    } else {
       const registerUser = await userRegistration(signUpData);
-    }
-  } catch (error) {
-    console.log(error.message);
-  }
 
-  res.send("Registration Successful");
+      res.send("User Created Sucessfully");
+    } catch (error) {
+      console.log(error.message);
+    }
+  }
 });
 
 // ? GET ALL USERS
 
-app.get("/users", authorisation, async (request, response) => {
-  const getDatas = request.headers;
-  const { token } = getDatas;
-  const crackData = await jsonwebtocken.verify(token, process.env.privateKey);
-  const { email } = crackData;
+app.get("/users", async (request, response) => {
+  const { token } = request.headers;
+  const { email } = await jsonwebtocken.verify(token, process.env.privateKey);
   const inputData = { email: email };
-  const checkUserExists = await findOneData(inputData);
-  if (!checkUserExists) {
+  const checkExists = await findOneData(inputData);
+  if (!checkExists) {
     response.status(401).send("Unauthorized user");
   } else {
     const getAllUserData = await getAllUsers();
@@ -155,39 +140,71 @@ app.get("/users", authorisation, async (request, response) => {
 
 app.put("/editUsers/:id", async (request, response) => {
   const { id } = request.params;
+
   const { name, contact, dob, email } = request.body;
 
-  const hashPassword = await createPassword("12345678");
+  const { token } = request.headers;
 
-  const hashConformPassword = await createPassword("12345678");
+  const crackData = await jsonwebtocken.verify(token, process.env.privateKey);
 
-  const updatedData = await {
-    name: name,
-    contact: contact,
-    dob: dob,
-    email: email,
-    hashPassword: hashPassword,
-    hashConformPassword: hashConformPassword,
-  };
-  const documentId = await new ObjectId(id);
-  const updateDatas = await modifyData(documentId, updatedData);
-  response.send(updateDatas);
+  const mail = crackData.email;
+
+  const inputData = { email: mail };
+
+  const checkUserExists = await findOneData(inputData);
+
+  if (!checkUserExists) {
+    response.status(401).send("Unauthorized user");
+  } else {
+    const hashPassword = await createPassword("12345678");
+
+    const hashConformPassword = await createPassword("12345678");
+
+    const updatedData = await {
+      name: name,
+      contact: contact,
+      dob: dob,
+      email: email,
+      hashPassword: hashPassword,
+      hashConformPassword: hashConformPassword,
+    };
+
+    const documentId = await new ObjectId(id);
+
+    const updateDatas = await modifyData(documentId, updatedData);
+
+    response.send(updateDatas);
+  }
 }),
   // ? DELETE TABLES DATA BY ID
 
   app.delete("/deleteUsers/:id", async (request, response) => {
     const { id } = request.params;
-    const hashPassword = await createPassword("12345678");
 
-    const hashConformPassword = await createPassword("12345678");
+    const { token } = request.headers;
 
-    const documentId = await new ObjectId(id);
-    const deleteDocId = { _id: documentId };
+    const crackData = await jsonwebtocken.verify(token, process.env.privateKey);
 
-    const deleteUserData = await deleteSingleUserData(deleteDocId);
-    response.send(deleteUserData);
-  }),
-  app.listen(PORT, () => console.log(`server started on port ${PORT} `));
+    const mail = crackData.email;
+
+    const inputData = { email: mail };
+
+    const checkUserExists = await findOneData(inputData);
+
+    if (!checkUserExists) {
+      response.status(401).send("Unauthorized user");
+    } else {
+      const documentId = await new ObjectId(id);
+
+      const deleteDocId = { _id: documentId };
+
+      const deleteUserData = await deleteSingleUserData(deleteDocId);
+
+      response.send(deleteUserData);
+    }
+  });
+
+app.listen(PORT, () => console.log(`server started on port ${PORT} `));
 
 //? DataBase Connection
